Edit Confidential Transfers tutorial.
Edits to Confidential Transfers tutorial, primarily in Steps 3 and 5.  In
Step 3, removed untrue or overstated claims about the privacy of the CT
feature, and replaced with warnings about what kinds of information can be
revealed by using blockchain analysis to "trace" the history of a blind
balance, and suggestions for safe use. In Step 5, made corrections to the
instructions for un-blinding a balance back to a public account.  Also added
a Summary section with brief review of the wallet commands covered in the
forgoing sections.